# Speed up scaling with gm