--- a/test/unit/utils/image/scaleImage.test.js
+++ b/test/unit/utils/image/scaleImage.test.js
@@ -12,6 +12,9 @@ const imagePath = path.join(process.cwd(), 'test', 'fixture', 'images');
 const imageBase = path.join(imagePath, 'base', 'base-image.png');
 const imageScaledDown = path.join(imagePath, 'scaled', 'scaled-down-image.png');
 
+const imageIphoneBase = path.join(imagePath, 'base', 'iOS_iPhone.png');
+const imageIphoneScaledDown = path.join(imagePath, 'scaled', 'iOS_iPhone_scaled_down.png');
+
 async function readAsBase64(file) {
     // read binary data
     const content = await fsExtra.readFile(file);
@@ -35,5 +38,18 @@ describe('scaleImage', function() {
 
     // then
     await compareImages(imageTest, imageScaledDown);
+  });
+
+  it('scales more complex image down', async function() {
+    // given
+    const base64ImageInput = await readAsBase64(imageIphoneBase);
+    const imageTest = path.join(tmpPath, 'iOS_iPhone_scaled_down.png');
+
+    // when
+    const base64ImageOutput = await scaleImage(base64ImageInput, 0.5);
+    await saveBase64Image(imageTest, base64ImageOutput);
+
+    // then
+    await compareImages(imageTest, imageIphoneScaledDown);
   });
 });