… test/fixture/images/base/iOS_iPhone.png
… test/fixture/images/scaled/iOS_iPhone_scaled_down.png