Previously scaling of screenshots was always done via Jimp. This PR enables scaling via GraphicsMagick when available.

Stats:

- url: https://github.com/zinserjan/wdio-screenshot
- method: saveDocumentScreenshot
- Measured before and after makeAreaScreenshot with console.time

| | Before #69 | After #69 | Now | Improvement (total)  |
| ------------- |:-------------:| -----:| -----:|-----:|
| iPad Air 2 iOS 10.2 landscape |  10825ms | 6888ms| 5881ms | 45% |
| Chrome v56 1024x768 | 6328ms | 5184ms | 2180ms | 65% |
| Firefox v51 1024x768 | 1890ms | 1485ms | 1477ms| no difference* |

* Firefox already emits non-retina images, that's why this PR does not have any impacts on the performance